chore(app-platform): align releasing, building, and testing
Align the way we build, and test, this library with the DHIS2 App
Platform standard using d2-app-scripts and d2-utils-cypress, along with
updating other tools like cli-style.

The biggest change is that this adds cli-utils-release to the process,
allowing this lib to be released in our standard way.

This means that merging to master triggers the release process.
