+++ .browserslistrc
@@ -1,5 +1,5 @@
 > 0.5%
 last 2 versions
 Firefox ESR
-ie 11
+not ie 11
 not dead
