+++ .dependabot/config.yml
@@ -1,7 +1,7 @@
 version: 1
 
 update_configs:
-  - package_manager: "javascript"
-    directory: "/"
-    update_schedule: "live"
-    version_requirement_updates: "increase_versions"
+    - package_manager: 'javascript'
+      directory: '/'
+      update_schedule: 'live'
+      version_requirement_updates: 'increase_versions'
