+++ .github/workflows/node-lint.yml
@@ -1,0 +1,26 @@
+name: 'DHIS2: Style'
+
+on:
+    push:
+        branches:
+            - master
+    pull_request:
+
+jobs:
+    pr:
+        name: Lint
+        runs-on: ubuntu-latest
+        if: "!contains(github.event.head_commit.message, '[skip ci]')"
+        steps:
+            - uses: actions/checkout@v1
+            - uses: actions/setup-node@v1
+              with:
+                  node-version: 12.x
+
+            - name: Install
+              run: yarn install --frozen-lockfile
+
+            - name: Run linters
+              run: yarn lint
+        env:
+            CI: true
