+++ .github/workflows/node-publish.yaml
@@ -1,0 +1,41 @@
+name: 'DHIS2: Release'
+
+on:
+    push:
+        branches:
+            - master
+
+jobs:
+    process:
+        name: Publish
+        runs-on: ubuntu-latest
+        if: "!contains(github.event.head_commit.message, '[skip ci]')"
+        steps:
+            - uses: actions/checkout@v1
+            - uses: actions/setup-node@v1
+              with:
+                  node-version: 12.x
+
+            - name: Install
+              run: yarn install --frozen-lockfile
+
+            - name: Lint
+              run: yarn lint
+
+            - name: Build
+              run: yarn build
+
+            - name: Test
+              run: yarn test
+
+            - name: Publish to NPM
+              run: npx @dhis2/cli-utils release --publish npm
+              env:
+                  GIT_AUTHOR_NAME: '@dhis2-bot'
+                  GIT_AUTHOR_EMAIL: 'apps@dhis2.org'
+                  GIT_COMMITTER_NAME: '@dhis2-bot'
+                  GIT_COMMITTER_EMAIL: 'apps@dhis2.org'
+                  NPM_TOKEN: ${{secrets.NPM_TOKEN}}
+                  GH_TOKEN: ${{secrets.GH_TOKEN}}
+        env:
+            CI: true
