+++ .huskyrc.js
@@ -1,6 +1,14 @@
+const { config } = require('@dhis2/cli-style')
+const husky = require(config.husky)
+
+const tasks = arr => arr.join(' && ')
+
 module.exports = {
     hooks: {
-        'commit-msg': 'd2-style commit check',
-        'pre-commit': 'd2-style validate',
+        ...husky.hooks,
+        'pre-commit': tasks([
+            'd2-style js check --staged',
+            'd2-style text check --staged',
+        ]),
     },
 }
