+++ .prettierrc.js
@@ -1,18 +1,5 @@
+const { config } = require('@dhis2/cli-style')
+
 module.exports = {
-    printWidth: 80,
-    tabWidth: 4,
-    useTabs: false,
-    semi: false,
-    singleQuote: true,
-    trailingComma: 'es5',
-    bracketSpacing: true,
-    jsxBracketSameLine: false,
-    jsxSingleQuote: false,
-    arrowParens: 'avoid',
-    rangeStart: 0,
-    rangeEnd: Infinity,
-    proseWrap: 'preserve',
-    requirePragma: false,
-    insertPragma: false,
-    endOfLine: 'lf',
+    ...require(config.prettier),
 }
