+++ cypress.json
@@ -1,5 +1,6 @@
 {
-  "testFiles": "**/*.feature",
-  "video": false,
-  "projectId": "konyyu"
+    "baseUrl": "http://localhost:5001",
+    "testFiles": "**/*.feature",
+    "video": false,
+    "projectId": "konyyu"
 }
