+++ cypress/support/index.js
@@ -1,3 +1,6 @@
+import '@dhis2/cli-utils-cypress/support'
+
+// Add additional support functions here
 import './getFormValue'
 import './screenshots'
 import './selectSelectNthOption'
