+++ jest.config.js
@@ -1,5 +1,3 @@
 module.exports = {
-    roots: ['./src'],
-    testRegex: '\\.test\\.js$',
     setupFilesAfterEnv: ['./src/setupTests.js'],
 }
