+++ package.json
@@ -1,5 +1,5 @@
 {
-    "name": "@dhis2/ui-form",
+    "name": "@dhis2/ui-forms",
     "version": "0.0.1",
     "main": "./build/cjs/index.js",
     "module": "./build/es/index.js",
@@ -13,62 +13,38 @@
         "access": "public"
     },
     "scripts": {
-        "extract-pot": "d2-i18n-extract -p src/ -o i18n/",
-        "generate-locale": "d2-i18n-generate -n 'ui-forms' -p ./i18n/ -o ./src/locales/",
-        "localize": "yarn extract-pot && yarn generate-locale",
-        "prebuild": "rm -rf ./build/* && yarn localize",
-        "build:commonjs": "BABEL_ENV=commonjs babel src --out-dir ./build/cjs --copy-files --verbose",
-        "build:modules": "BABEL_ENV=modules babel src --out-dir ./build/es --copy-files --verbose",
-        "build": "NODE_ENV=production yarn build:commonjs && NODE_ENV=production yarn build:modules",
-        "postbuild": "NODE_ENV=production yarn build-storybook",
-        "build-storybook": "build-storybook -c .storybook -o ./build/docs",
-        "start-storybook": "start-storybook --port 5000",
-        "prestart": "yarn localize",
-        "start": "yarn start-storybook",
-        "start:testing": "STORYBOOK_INCLUDE_TESTING=1 yarn start --ci --port $PORT",
-        "cypress:browser": "cypress open --config baseUrl=http://localhost:$PORT",
-        "cypress:ci": "NO_COLOR=1 cypress run -b $BROWSER --config baseUrl=http://localhost:$PORT",
-        "cypress:open": "PORT=12345 concurrently --kill-others -n storybook,cypress 'yarn start:testing' 'wait-on http-get://localhost:$PORT && yarn cypress:browser'",
-        "cypress:run": "PORT=12345 concurrently --kill-others --success first -n storybook,cypress 'yarn start:testing' 'wait-on http-get://localhost:$PORT && yarn cypress:ci'",
-        "test": "jest"
-    },
-    "husky": {
-        "hooks": {
-            "commit-msg": "d2-style commit check",
-            "pre-commit": "yarn test && d2-style js apply"
-        }
-    },
-    "cypress-cucumber-preprocessor": {
-        "nonGlobalStepDefinitions": true
+        "build": "d2-app-scripts build",
+        "build:docs": "d2-utils-docsite build ./docs -o ./build/docs --jsdoc ./src --jsdoc-output-file api.md",
+        "build:demo": "EXTEND_ESLINT=true build-storybook -c .storybook -o ./build/docs/demo",
+        "start": "yarn start:demo",
+        "start:docs": "d2-utils-docsite serve ./docs -o ./build/docs --jsdoc ./src --jsdoc-output-file api.md",
+        "start:demo": "EXTEND_ESLINT=true start-storybook --port 5000",
+        "start:testing": "EXTEND_ESLINT=true STORYBOOK_INCLUDE_TESTING=1 start-storybook --port 5001 --quiet --ci",
+        "cy:open": "concurrently --kill-others -n storybook,cypress 'yarn start:testing' 'wait-on http-get://localhost:5001 && d2-utils-cypress open'",
+        "cy:run": "concurrently --kill-others --success first -n storybook,cypress 'yarn start:testing' 'wait-on http-get://localhost:5001 && d2-utils-cypress run'",
+        "lint": "yarn lint:js && yarn lint:text",
+        "lint:js": "d2-style js check",
+        "lint:text": "d2-style text check",
+        "format": "yarn format:js && yarn format:text",
+        "format:staged": "yarn format:js --staged && yarn format:text --staged",
+        "format:js": "d2-style js apply",
+        "format:text": "d2-style text apply",
+        "test": "d2-app-scripts test"
     },
     "devDependencies": {
-        "@babel/cli": "^7.8.3",
-        "@babel/core": "^7.8.3",
-        "@babel/plugin-proposal-class-properties": "^7.8.3",
-        "@babel/plugin-proposal-object-rest-spread": "^7.8.0",
-        "@babel/plugin-transform-react-constant-elements": "^7.8.3",
-        "@babel/preset-env": "^7.8.3",
-        "@babel/preset-react": "^7.8.3",
         "@dhis2/app-runtime": "2.0.4",
-        "@dhis2/cli-style": "5.1.0",
-        "@dhis2/d2-i18n-extract": "1.0.8",
-        "@dhis2/d2-i18n-generate": "1.1.1",
-        "@dhis2/d2-i18n": "1.0.6",
+        "@dhis2/cli-app-scripts": "^3.2.3",
+        "@dhis2/cli-style": "^6.0.0",
+        "@dhis2/cli-utils-cypress": "^1.0.2",
         "@dhis2/ui-core": "4.1.2",
         "@storybook/addon-notes": "5.3.8",
+        "@storybook/preset-create-react-app": "^1.5.2",
         "@storybook/react": "^5.3.8",
         "@storybook/router": "^5.3.8",
-        "babel-eslint": "10.x",
-        "babel-loader": "^8.0.5",
         "concurrently": "^5.0.2",
-        "cypress": "^3.8.2",
-        "cypress-cucumber-preprocessor": "^2.0.1",
         "enzyme": "^3.11.0",
         "enzyme-adapter-react-16": "^1.15.2",
-        "eslint-plugin-react": "^7.18.0",
         "fs-extra": "^8.1.0",
-        "husky": "^4.1.0",
-        "jest": "^24.9.0",
         "react": "^16.9.0",
         "react-dev-utils": "^10.0.0",
         "react-dom": "^16.9.0",
