--- release.config.js
@@ -1,4 +1,0 @@
-module.exports = {
-    branch: 'master',
-    tagFormat: 'v${version}',
-}
