+++ src/__demo__/SingleSelect.stories.js
@@ -1,8 +1,7 @@
 import React from 'react'
 import { storiesOf } from '@storybook/react'
 
-import { Field, MultiSelect } from '../src/index.js'
-import { formDecorator } from '../.storybook/formDecorator.js'
+import { Field, SingleSelect } from '../index.js'
 
 const options = [
     { value: '1', label: 'one' },
@@ -17,18 +16,12 @@ const options = [
     { value: '10', label: 'ten' },
 ]
 
-const initialValue = [
-    { value: '3', label: 'three' },
-    { value: '4', label: 'four' },
-    { value: '9', label: 'nine' },
-    { value: '10', label: 'ten' },
-]
+const initialValue = { value: '4', label: 'four' }
 
-storiesOf('MultiSelect', module)
-    .addDecorator(formDecorator)
+storiesOf('SingleSelect', module)
     .add('Default', () => (
         <Field
-            component={MultiSelect}
+            component={SingleSelect}
             name="agree"
             label="Do you agree?"
             options={options}
@@ -36,7 +29,7 @@ storiesOf('MultiSelect', module)
     ))
     .add('InitialValue', () => (
         <Field
-            component={MultiSelect}
+            component={SingleSelect}
             name="agree"
             label="Do you agree?"
             options={options}
