+++ src/__e2e__/Checkbox.stories.js
@@ -1,16 +1,14 @@
 import React from 'react'
 import { storiesOf } from '@storybook/react'
 
-import { Field, Switch, hasValue } from '../src/index.js'
-import { formDecorator } from '../.storybook/formDecorator.js'
+import { Field, Checkbox, hasValue } from '../index.js'
 
-storiesOf('Testing:Switch', module)
-    .addDecorator(formDecorator)
+storiesOf('Testing:Checkbox', module)
     .add('Unchecked', () => (
         <Field
-            component={Switch}
-            className="switch"
-            name="switch"
+            component={Checkbox}
+            className="checkbox"
+            name="checkbox"
             label="Label text"
             validate={hasValue}
             required
@@ -18,27 +16,27 @@ storiesOf('Testing:Switch', module)
     ))
     .add('Checked ', () => (
         <Field
-            component={Switch}
-            className="switch"
-            name="switch"
+            component={Checkbox}
+            className="checkbox"
+            name="checkbox"
             label="Label text"
             initialValue={true}
         />
     ))
     .add('Unchecked with value', () => (
         <Field
-            component={Switch}
-            className="switch"
-            name="switch"
+            component={Checkbox}
+            className="checkbox"
+            name="checkbox"
             label="Label text"
             checkedValue="yes"
         />
     ))
     .add('Checked with value', () => (
         <Field
-            component={Switch}
-            className="switch"
-            name="switch"
+            component={Checkbox}
+            className="checkbox"
+            name="checkbox"
             label="Label text"
             checkedValue="yes"
             initialValue="yes"
