+++ src/__e2e__/MultiSelect.stories.js
@@ -1,8 +1,7 @@
 import { Field } from 'react-final-form'
 import React from 'react'
 import { storiesOf } from '@storybook/react'
-import { MultiSelect, hasValue } from '../src/index.js'
-import { formDecorator } from '../.storybook/formDecorator.js'
+import { MultiSelect, hasValue } from '../index.js'
 
 const defaultOptions = [
     { value: 'first', label: 'First' },
@@ -10,7 +9,6 @@ const defaultOptions = [
 ]
 
 storiesOf('Testing:MultiSelect', module)
-    .addDecorator(formDecorator)
     .addParameters({ options: { showPanel: false } })
     .add('Required', ({ cypressProps }) => (
         <Field
