+++ src/__e2e__/SingleSelect.stories.js
@@ -1,13 +1,11 @@
 import { Field } from 'react-final-form'
 import React from 'react'
 import { storiesOf } from '@storybook/react'
-import { SingleSelect, hasValue } from '../src/index.js'
-import { formDecorator } from '../.storybook/formDecorator.js'
+import { SingleSelect, hasValue } from '../index.js'
 
 const defaultOptions = [{ value: 'initial', label: 'Initial' }]
 
 storiesOf('Testing:SingleSelect', module)
-    .addDecorator(formDecorator)
     .addParameters({ options: { showPanel: false } })
     .add('Required', ({ cypressProps }) => (
         <Field
