+++ src/__e2e__/Switch.stories.js
@@ -1,16 +1,14 @@
 import React from 'react'
 import { storiesOf } from '@storybook/react'
 
-import { Field, Checkbox, hasValue } from '../src/index.js'
-import { formDecorator } from '../.storybook/formDecorator.js'
+import { Field, Switch, hasValue } from '../index.js'
 
-storiesOf('Testing:Checkbox', module)
-    .addDecorator(formDecorator)
+storiesOf('Testing:Switch', module)
     .add('Unchecked', () => (
         <Field
-            component={Checkbox}
-            className="checkbox"
-            name="checkbox"
+            component={Switch}
+            className="switch"
+            name="switch"
             label="Label text"
             validate={hasValue}
             required
@@ -18,27 +16,27 @@ storiesOf('Testing:Checkbox', module)
     ))
     .add('Checked ', () => (
         <Field
-            component={Checkbox}
-            className="checkbox"
-            name="checkbox"
+            component={Switch}
+            className="switch"
+            name="switch"
             label="Label text"
             initialValue={true}
         />
     ))
     .add('Unchecked with value', () => (
         <Field
-            component={Checkbox}
-            className="checkbox"
-            name="checkbox"
+            component={Switch}
+            className="switch"
+            name="switch"
             label="Label text"
             checkedValue="yes"
         />
     ))
     .add('Checked with value', () => (
         <Field
-            component={Checkbox}
-            className="checkbox"
-            name="checkbox"
+            component={Switch}
+            className="switch"
+            name="switch"
             label="Label text"
             checkedValue="yes"
             initialValue="yes"
