+++ src/components/FormSpy.js
@@ -4,9 +4,9 @@ import propTypes from '@dhis2/prop-types'
 FormSpy.propTypes = {
     children: propTypes.func,
     component: propTypes.func,
-    onChange: propTypes.func,
     render: propTypes.func,
     subscription: propTypes.object,
+    onChange: propTypes.func,
 }
 
 export { FormSpy }
